# docs/openapi.yaml
...
components:
...
    schemas:
        ActivateRewardResponse:
            properties:
                code:
                    type: string
                reward_id:
                    type: string
                status:
                    type: string
            type: object
        Badge:
            properties:
                description:
                    type: string
                earned_at:
                    type: string
                icon_url:
                    type: string
                id:
                    type: string
                name:
                    type: string
            type: object
...
        ClaimRewardRequest:
            properties:
                reward_id:
                    type: string
            required:
                - reward_id
            type: object
        ClaimRewardResponse:
            properties:
                promocode:
                    type: string
                reward_id:
                    type: string
            type: object
...
        ErrorBody:
            properties:
                code:
                    type: string
                field:
                    type: string
                message:
                    type: string
                request_id:
                    type: string
            type: object
        ErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ErrorBody'
            type: object
...
        Pet:
            properties:
                energy:
                    type: integer
                freezes:
                    type: integer
                happiness:
                    type: integer
                hatched_at:
                    type: string
                id:
                    type: string
                is_hatched:
                    type: boolean
                last_checkin_date:
                    type: string
                last_decay_time:
                    type: string
                level:
                    type: integer
                name:
                    type: string
                next_level_xp:
                    type: integer
                satiety:
                    type: integer
                stage:
                    $ref: '#/components/schemas/PetStage'
                state:
                    $ref: '#/components/schemas/PetState'
                streak_days:
                    type: integer
                updated_at:
                    type: string
                user_id:
                    type: string
                xp:
                    type: integer
            type: object
        PetStage:
            enum:
                - egg
                - baby
                - teen
                - adult
                - legend
            type: string
            x-enum-varnames:
                - StageEgg
                - StageBaby
                - StageTeen
                - StageAdult
                - StageLegend
        PetState:
            enum:
                - happy
                - neutral
                - sad
                - sleeping
            type: string
            x-enum-varnames:
                - StateHappy
                - StateNeutral
                - StateSad
                - StateSleeping
...
        RaccoonProfile:
            properties:
                badges:
                    items:
                        $ref: '#/components/schemas/Badge'
                    type: array
                    uniqueItems: false
                current_streak:
                    type: integer
                id:
                    type: string
                level:
                    type: integer
                name:
                    type: string
                stage:
                    type: string
                state:
                    type: string
                user_id:
                    type: string
                xp:
                    type: integer
                xp_to_next_level:
                    type: integer
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            description: |
                JWT, подписанный HS256. Клеймы: `sub` — UUID пользователя,
                `role` — роль, `exp` — момент истечения, `iat` — момент выпуска,
                `jti` — уникальный идентификатор токена.

                Просроченный токен и токен с неверной подписью одинаково
                дают 401 с кодом `unauthorized`.
            in: header
            name: Authorization
            scheme: bearer
            type: http
info:
    contact:
        name: Avito Hack Team
    description: |
        API объявлений с игровым слоем: пользователь заводит объявления, добавляет фотографии,
        ведёт избранное, а параллельно развивает питомца — за действия в маркетплейсе
        начисляется опыт, растут уровень и стрик, открываются награды.

        ## Конверт ошибок

        Все ошибки возвращаются единым конвертом `ErrorEnvelope`:

        ```json
        {"error": {"code": "validation_error", "message": "title must be at least 3 characters", "field": "title", "request_id": "..."}}
        ```

        Поля `field` и `request_id` опциональны и опускаются, когда пусты.
        Коды: `validation_error`, `bad_request`, `unauthorized`, `forbidden`,
        `not_found`, `conflict`, `internal_error`.

        ## Авторизация

        JWT (HS256) передаётся в заголовке `Authorization: Bearer <token>`.
        Срок жизни токена лежит в клейме `exp` внутри самого токена — отдельного поля
        `expires_at` в ответах на регистрацию и вход нет. В клеймах также присутствуют
        `sub` (идентификатор пользователя) и `role` (`user` | `moderator` | `admin`).
        WebSocket авторизуется query-параметром `token`, так как браузерный
        WebSocket API не позволяет задавать заголовки.

        ## Пагинация

        Списки используют keyset-курсоры, непрозрачные для клиента (base64url).
        В проекте два независимых формата курсора:

...

...

        ## Деньги

...

        ## Тело запроса

...
    license:
...
openapi: 3.1.0
paths:
...
    /api/v1/badges:
        get:
            description: |-
                Плоский массив бейджей — без обёртки `{items}` и без пагинации.
                Поле `earned_at` пусто (и опущено) у ещё не полученных бейджей;
                формат времени — RFC 3339 в UTC.
            operationId: listBadges
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Badge'
                                type: array
                    description: Массив бейджей
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Бейджи пользователя
            tags:
                - Raccoon
...
    /api/v1/pet:
        get:
            description: |-
                Текущее состояние питомца владельца токена. Питомец создаётся
                автоматически при регистрации, поэтому 404 здесь не ожидается.

                Побочный эффект чтения: перед выдачей применяется ленивое затухание
                характеристик (`satiety`, `happiness`, `energy`) пропорционально времени,
                прошедшему с `last_decay_time`. Поэтому два последовательных чтения
                без действий пользователя могут вернуть разные значения характеристик
                и производное поле `state`.
            operationId: getPet
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Pet'
                    description: Состояние питомца
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Состояние питомца
            tags:
                - Pet
...
    /api/v1/raccoon/profile:
        get:
            description: |-
                Альтернативное представление того же питомца, ориентированное на
                экран профиля: уровень, опыт, остаток до следующего уровня, стрик,
                стадия, состояние и заработанные бейджи.

                Данные берутся из того же сервиса, что и `GET /api/v1/pet`, но поля
                названы иначе — `xp_to_next_level` вместо `next_level_xp`,
                `current_streak` вместо `streak_days`. Все идентификаторы здесь
                отдаются строками, хотя по сути являются UUID.
            operationId: getRaccoonProfile
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RaccoonProfile'
                    description: Профиль енота
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Профиль енота
            tags:
                - Raccoon
...
    /api/v1/rewards/{id}/activate:
        post:
            description: |-
                Превращает выданную награду в промокод. Код детерминированно
                выводится через HMAC от идентификаторов пользователя и награды,
                поэтому одна и та же награда всегда даёт один и тот же код.

                Не идемпотентна на уровне ответа: первая активация даёт 200,
                повторная — 409 (`reward has already been activated`).

                Ошибки:

                - награда не выдана пользователю — 400, `field: reward_id`;
                - условие награды не выполнено — 400, `field: reward_id`;
                - награда уже активирована — 409;
                - награда просрочена — 409;
                - награда неактивируемая по своей природе — 409.
            operationId: activateReward
            parameters:
                - description: Строковый идентификатор награды (например `streak_5`), НЕ UUID.
                  in: path
                  name: id
                  required: true
                  schema:
                    minLength: 1
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ActivateRewardResponse'
                    description: Награда активирована
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Награда не выдана или условие не выполнено
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Награда уже активирована, просрочена или неактивируема
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Активация награды
            tags:
                - Rewards
    /api/v1/rewards/claim:
        post:
            description: |-
                Вариант активации награды из модуля енота: идентификатор награды
                передаётся в теле, а не в пути, и в ответе поле называется
                `promocode`, а не `code`.

                Функционально дублирует `POST /api/v1/rewards/{id}/activate`
                и работает поверх того же сервиса наград. Для нового кода
                предпочтительна версия с идентификатором в пути.

                Не идемпотентна: повторный вызов для уже активированной награды
                даёт 409.
            operationId: claimReward
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ClaimRewardRequest'
                description: Идентификатор награды
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ClaimRewardResponse'
                    description: Промокод выдан
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Conflict
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Получить промокод награды
            tags:
                - Raccoon
...
servers:
...